CAT-949 Remove provenance section, convert donor page to TypeScript
